Revert "Added a main method to TestUtilities to populate the database"
This reverts commit 2e9b2a5154fb4e224d00583738ded10bfafdb2dd.

This was done because Lane's commit was extremely **broken** and I was
sick of wasting time trying to fix all the messes it caused.  Here are
the main ones:
- It failed to handle THREE checked exceptions, breaking the build of
  all test classes.  This alone serves as clear evidence that //no
  attempt had ever been made to compile the code before committing it//.
- It somehow managed to inadvertently (?) revert Matt's preceding
  commit, re-introducing a regression that, when interacting with the
  build system, makes a failed test appear in every run.
Please fix this *properly* when you're willing to put in the time to
make sure it actually compiles and runs before you commit it and waste
others' time.