Wrote VendingMachine tests and debugged swapInNextLayout()
That method was guilty of overwriting the remaining quantities of rows
that were in the same location but selling different products in the new
layout than they had in the old.  Woah, Nilly!  Or, as my high school
physics teacher would say, "Cool it, hot shot!"  And cool it I have.