Modified handling of believed-to-be impossible exceptions
Although these used to be silently consumed by the model, they are now
consumed and announced to standard error.  In this way, higher levels
don't have to deal with utterly unexpected situations (read:
intra-model implementation contradictions), but the operator is still
able to see that "something" is going quite wrong.